requirements.txt
aiohttp>=3.8.0
pandas>=2.0.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0
pyarrow>=14.0.1
numpy>=1.24.0
boto3>=1.28.0
botocore>=1.31.0
aiodns>=3.0.0  # Optional: For better DNS resolution performance
cchardet>=2.1.7  # Optional: For better character encoding detection
ujson>=5.7.0  # Optional: For faster JSON processing
aiofiles>=23.1.0  # For async file operations
cryptography>=41.0.0  # For secure connections
tenacity>=8.2.0  # For better retry mechanisms
websockets>=12.0  # For instrument fetching
pyyaml>=6.0.1  # For configuration loading
pytest>=7.0.0  # For testing
pytest-asyncio>=0.23.0  # For async testing support
seaborn>=0.13.0  # For analytics visualization
matplotlib>=3.8.0  # For analytics plotting
